[CDEC-344] Refactor Pos.Core.Slotting module set
