+++ core/src/Pos/Aeson/Core.hs
@@ -28,8 +28,8 @@ import           Pos.Core.Common (Address, BlockCount (..), ChainDifficulty, Coi
                                   decodeTextAddress, mkCoin, unsafeCoinPortionFromDouble,
                                   unsafeGetCoin)
 import           Pos.Core.Delegation (HeavyDlgIndex (..))
-import           Pos.Core.Slotting.Types (EpochIndex (..), LocalSlotIndex, SlotCount (..), SlotId,
-                                          Timestamp (..))
+import           Pos.Core.Slotting (EpochIndex (..), LocalSlotIndex, SlotCount (..), SlotId,
+                                    Timestamp (..))
 import           Pos.Core.Ssc (VssCertificate)
 import           Pos.Core.Update.Types (ApplicationName (..), BlockVersion, BlockVersionData,
                                         SoftforkRule, SoftwareVersion (..))
