+++ core/src/Pos/Binary/Core/Update.hs
@@ -18,7 +18,7 @@ import           Pos.Binary.Core.Fee ()
 import           Pos.Binary.Core.Script ()
 import           Pos.Binary.Core.Slotting ()
 import           Pos.Core.Common (CoinPortion, ScriptVersion, TxFeePolicy)
-import           Pos.Core.Slotting.Types (EpochIndex, FlatSlotId)
+import           Pos.Core.Slotting (EpochIndex, FlatSlotId)
 import           Pos.Core.Update.Types (ApplicationName (..), BlockVersion (..),
                                         BlockVersionData (..), BlockVersionModifier (..),
                                         NumSoftwareVersion, SoftforkRule (..), SoftwareVersion (..),
