+++ core/src/Pos/Core/Configuration.hs
@@ -30,10 +30,10 @@ import           Pos.Core.Genesis.Canonical (SchemaError)
 import           Pos.Core.Genesis.Generate (GeneratedGenesisData (..), generateGenesisData)
 import           Pos.Core.Genesis.Helpers (mkGenesisDelegation)
 import           Pos.Core.Genesis.Types (GenesisData (..), GenesisDelegation,
-                                         GenesisInitializer (..), GenesisSpec (..),
-                                         GenesisProtocolConstants (..),
+                                         GenesisInitializer (..), GenesisProtocolConstants (..),
+                                         GenesisSpec (..),
                                          genesisProtocolConstantsToProtocolConstants)
-import           Pos.Core.Slotting.Types (Timestamp)
+import           Pos.Core.Slotting (Timestamp)
 import           Pos.Crypto.Configuration as E
 import           Pos.Crypto.Hashing (Hash, hashRaw, unsafeHash)
 import           Pos.Util.Util (leftToPanic)
