+++ core/src/Pos/Core/Configuration/Protocol.hs
@@ -19,9 +19,8 @@ import           Universum
 import           Data.Reflection (Given (..), give)
 
 import           Pos.Core.Common (BlockCount (..))
-import           Pos.Core.ProtocolConstants (ProtocolConstants (..), VssMaxTTL (..),
-                                             VssMinTTL (..))
-import           Pos.Core.Slotting.Types (SlotCount)
+import           Pos.Core.ProtocolConstants (ProtocolConstants (..), VssMaxTTL (..), VssMinTTL (..))
+import           Pos.Core.Slotting.SlotCount (SlotCount)
 
 type HasProtocolConstants = Given ProtocolConstants
