+++ core/src/Pos/Core/Genesis/Canonical.hs
@@ -39,7 +39,7 @@ import           Pos.Core.Genesis.Types (GenesisAvvmBalances (..), GenesisData (
                                          GenesisProtocolConstants (..),
                                          GenesisVssCertificatesMap (..), GenesisWStakeholders (..))
 import           Pos.Core.ProtocolConstants (VssMaxTTL (..), VssMinTTL (..))
-import           Pos.Core.Slotting.Types (EpochIndex (..), Timestamp (..))
+import           Pos.Core.Slotting (EpochIndex (..), Timestamp (..))
 import           Pos.Core.Ssc (VssCertificate (..), VssCertificatesMap (..),
                                validateVssCertificatesMap)
 import           Pos.Core.Update.Types (BlockVersionData (..), SoftforkRule (..))
