+++ core/src/Pos/Core/Genesis/Types.hs
@@ -39,7 +39,7 @@ import           Pos.Binary.Crypto ()
 import           Pos.Core.Common (Address, Coin, CoinPortion, SharedSeed, StakeholderId)
 import           Pos.Core.Delegation (ProxySKHeavy)
 import           Pos.Core.ProtocolConstants (ProtocolConstants (..), VssMaxTTL (..), VssMinTTL (..))
-import           Pos.Core.Slotting.Types (Timestamp)
+import           Pos.Core.Slotting (Timestamp)
 import           Pos.Core.Ssc (VssCertificatesMap, getVssCertificatesMap)
 import           Pos.Core.Update.Types (BlockVersionData)
 import           Pos.Crypto.Configuration (ProtocolMagic)
