+++ core/src/Pos/Core/Slotting.hs
@@ -1,10 +1,17 @@
 module Pos.Core.Slotting
-       ( module Pos.Core.Slotting.Timestamp
-       , module Pos.Core.Slotting.Types
-       , module Pos.Core.Slotting.Util
+       ( module Pos.Core.Slotting.EpochIndex
+       , module Pos.Core.Slotting.EpochOrSlot
+       , module Pos.Core.Slotting.LocalSlotIndex
+       , module Pos.Core.Slotting.SlotCount
+       , module Pos.Core.Slotting.SlotId
+       , module Pos.Core.Slotting.TimeDiff
+       , module Pos.Core.Slotting.Timestamp
        ) where
 
+import           Pos.Core.Slotting.EpochIndex
+import           Pos.Core.Slotting.EpochOrSlot
+import           Pos.Core.Slotting.LocalSlotIndex
+import           Pos.Core.Slotting.SlotCount
+import           Pos.Core.Slotting.SlotId
+import           Pos.Core.Slotting.TimeDiff
 import           Pos.Core.Slotting.Timestamp
-import           Pos.Core.Slotting.Types
-import           Pos.Core.Slotting.Util
-
