+++ core/src/Pos/Core/Ssc/VssCertificatesMap.hs
@@ -27,7 +27,7 @@ import           Serokell.Util (allDistinct)
 
 import           Pos.Binary.Class (Bi)
 import           Pos.Core.Common (StakeholderId)
-import           Pos.Core.Slotting.Types (EpochIndex)
+import           Pos.Core.Slotting (EpochIndex)
 import           Pos.Crypto (ProtocolMagic)
 
 import           Pos.Core.Ssc.VssCertificate (VssCertificate (..), checkVssCertificate, getCertId,
